Refactor tagsBar: extract inner TagButton component
