+++ src/components/molecules/item/list/tagsBar/tagsBar.stories.jsx
@@ -3,11 +3,11 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 
-import theme from '../../../../styles/global-styles';
-import themeDecorator from '../../../../../.storybook/decorators/themeDecorator';
-import tags from '../../../../../.storybook/fake_data/tags.json';
+import theme from '../../../../../styles/global-styles';
+import themeDecorator from '../../../../../../.storybook/decorators/themeDecorator';
+import tags from '../../../../../../.storybook/fake_data/tags.json';
 
-import Component from './tagsBar';
+import Component from '.';
 
 storiesOf('Molecules/[Item]/List/TagsBar', module)
   .addDecorator(themeDecorator(theme))
